refactor: Move user profiles to bottom of sidenav
